==> 1. Notebooks/01_pobieranie_i_przetwarzanie_danych.ipynb <==
### Główne założenia:
#### Projekt bazuje na informacjach udostępnionych na stronie https://openweathermap.org/api przez firmę **OpenWeather**.

Na potrzeby projektu, z uwzględnieniem aspektów finansowych, zdecydowano się nie łączyć bezpośrednio z API strony, lecz zakupiono dane historyczne z okresu: ***01.01.1971 do 31.12.2023*** dla trzech przykładowych lokalizacji:

-   **Bożańska**, woj. pomorskie, gmina Szemud, Polska
-   **Wałcz Drugi**, woj. zachodniopomorskie, gmina Wałcz, Polska
-   **Baden-Baden**, Baden-Württemberg, Niemcy

### Parametry danych pogodowych:

- **city_name** - Nazwa miasta
- **lat** - Geograficzne współrzędne lokalizacji (szerokość geograficzna)
- **lon** - Geograficzne współrzędne lokalizacji (długość geograficzna)
- **main**:
    - **main.temp** - Temperatura [Celsjusz]
    - **main.temp_min** - Minimalna temperatura w danym momencie. Jest to odchylenie od temperatury możliwej dla dużych miast i megalopolis geograficznie rozszerzonych (użyj tego parametru opcjonalnie) [Celsjusz].
    - **main.temp_max** - Maksymalna temperatura w danym momencie. Jest to odchylenie od temperatury możliwej dla dużych miast i megalopolis geograficznie rozszerzonych (użyj tego parametru opcjonalnie) [Celsjusz].
    - **main.feels_like** - Ten parametr temperatury uwzględnia ludzkie odczucie pogody [Celsjusz]
    - **main.pressure** - Ciśnienie atmosferyczne (na poziomie morza), [hPa]
    - **main.humidity** - Wilgotność, [%]
    - **main.dew_point** - Temperatura atmosferyczna (zmieniająca się zgodnie z ciśnieniem i wilgotnością), poniżej której krople wody zaczynają kondensować się i tworzy się rosa [Celsjusz]
- **wind**:
    - **wind.speed** - Prędkość wiatru [metr/sek]
    - **wind.deg** - Kierunek wiatru, [stopnie] (meteorologiczne)
    - **wind.gust** - Poryw wiatru [metr/sek]
- **clouds**:
    - **clouds.all** - Zachmurzenie, [%]
- **rain**:
    - **rain.1h** - Ilość opadów deszczu za ostatnią godzinę, [mm]
    - **rain.3h** - Ilość opadów deszczu za ostatnie 3 godziny, [mm]
- **snow**:
    - **snow.1h** - Ilość opadów śniegu za ostatnią godzinę, [mm] (w stanie ciekłym)
    - **snow.3h** - Ilość opadów śniegu za ostatnie 3 godziny, [mm] (w stanie ciekłym)
- **weather**:
    - **weather.id** - Identyfikator warunków pogodowych [-]
    - **weather.main** - Grupa parametrów pogodowych (Deszcz, Śnieg, Ekstremalne, itp.) [-]
    - **weather.description** - Opis warunków pogodowych w grupie. Więcej informacji znajdziesz tutaj [-]
    - **weather.icon** - Identyfikator ikony pogodowej [-]
- **visibility** - Średnia widoczność, metry. Maksymalna wartość widoczności to 10 km [m]
- **dt** - Czas obliczenia danych, unix, UTC
- **dt_iso** - Data i czas w formacie UTC
- **timezone** - Przesunięcie w sekundach względem UTC


# Importowanie niezbędnych bibliotek
import pandas as pd  # Biblioteka do manipulacji i analizy danych
import os  # Biblioteka do operacji na systemie plików, takich jak zarządzanie ścieżkami i plikami

<u>*_Uwaga:_*</u>

*Ten skrypt najpierw tworzy listę wszystkich plików CSV w danym folderze, a następnie ładuje każdy z nich do osobnej ramki danych. Dodatkowo, przykład pokazuje, jak można dodać nazwę miasta do każdego zestawu danych na podstawie nazwy pliku. Po przetworzeniu wszystkich plików, ramki danych są łączone w jedną dużą ramkę, którą można następnie zapisać lub dalej przetwarzać.*

# Ścieżka do folderu zawierającego pliki CSV
folder_path = '../2. Data/raw/'

# Lista plików CSV w folderze, filtrując tylko te, które mają rozszerzenie .csv
csv_files = [file for file in os.listdir(folder_path) if file.endswith('.csv')]

# Lista do przechowywania danych z każdego pliku
all_data = []

# Iteracja przez każdy plik w liście plików CSV
for file in csv_files:
    # Pełna ścieżka do pliku
    file_path = os.path.join(folder_path, file)
    
    # Wczytanie danych z pliku
    df = pd.read_csv(file_path)
    
    # Usunięcie rozszerzenia .csv z nazwy miasta i dodanie jako kolumna
    city_name = file.replace('.csv', '')
    df['city_name'] = city_name
    
    # Dodanie DataFrame do listy
    all_data.append(df)

# Połączenie wszystkich DataFrame w jedną
combined_data = pd.concat(all_data, ignore_index=True)

# Zapisanie połączonych danych do nowego pliku CSV z kompresją gzip
combined_data.to_csv('../2. Data/processed/combined_data.csv', index=False, compression='gzip')


# Wczytanie skompresowanego pliku CSV do DataFrame
df_weather = pd.read_csv("../2. Data/processed/combined_data.csv", compression='gzip')

# Wyświetlenie pierwszych pięciu wierszy danych z DataFrame df_weather
df_weather.head(5)

# Pobranie unikalnych nazw miast z kolumny "city_name" w DataFrame df_weather
df_weather["city_name"].unique()

# Sprawdzenie ogólnych informacji na temat danych w DataFrame df_weather
df_weather.info()

**int64:** Oznacza 64-bitową liczbę całkowitą. Jest to typ danych przechowujący liczby całkowite, które mogą być bardzo duże, ponieważ są reprezentowane przez 64 bity.

**object:** Oznacza typ ogólny, który w pandas często jest używany do przechowywania tekstu (ciągów znaków). Może także przechowywać inne typy danych, ale najczęściej jest używany do tekstu.

**float64:** Oznacza 64-bitową liczbę zmiennoprzecinkową. Jest to typ danych przechowujący liczby zmiennoprzecinkowe, które mogą mieć wiele miejsc po przecinku, ponieważ są reprezentowane przez 64 bity.

# Obliczenie liczby brakujących danych dla każdej kolumny w DataFrame df_weather
missing_data = df_weather.isnull().sum()

# Wyświetlenie wyników
missing_data

#### Z przedstawionego zestawienia dla DataFrame df_weather wynika, że liczba brakujących danych w poszczególnych kolumnach jest następująca:

- **visibility**: Brakuje danych w wierszach.
- **sea_level**: Wszystkie danych są puste.
- **grnd_level**: Wszystkie danych są puste.
- **wind_gust**: Brakuje danych w wierszach.
- **rain_1h**: Brakuje danych w wierszach.
- **rain_3h**: Wszystkie danych są puste.
- **snow_1h**: Brakuje danych w wierszach.
- **snow_3h**: Wszystkie danych są puste.

Pozostałe kolumny mają pełne zestawienie danych bez braków.

Na tej podstawie można podjąć kilka działań w celu przygotowania zestawu danych do analizy:

- Kolumny **sea_level**, **grnd_level**, **rain_3h**, **snow_3h**: Ponieważ wszystkie dane w tych kolumnach są puste, najprawdopodobniej najlepszym rozwiązaniem będzie ich usunięcie, nie jest planowane ich wypełnienie danymi z innego źródła na tym etapie projektu.

- Kolumny z dużą ilością brakujących danych **wind_gust**, **rain_1h**, **snow_1h**: Jeżeli te dane są kluczowe dla analizy, można rozważyć ich uzupełnienie, na przykład przez interpolację, użycie średnich z innych kolumn, lub całkowite usunięcie wierszy, w których brak danych. Na tym etapie projektu została podjęta decyzja o usunięciu tych kolumn.

- Kolumna **visibility**: Możliwe, że brakujące dane można uzupełnić, jeśli mają one wpływ na przeprowadzaną analizę, lub można te wiersze usunąć, jeżeli nie są one kluczowe. Na tym etapie projektu została podjęta decyzja po uzupełnieniu brakujących danych. Do dyspozycji są następujące metody:

  - **Imputacja średnią**: Jeżeli brakujące dane są przypadkowe i nie mają wyraźnego wzorca, możliwym rozwiązaniem jest imputacja, czyli wypełnienie brakujących wartości średnią lub medianą z dostępnych danych dla visibility. W przypadku bardziej złożonych danych można zastosować modelowanie predykcyjne (na przykład regresję liniową) bazujące na innych zmiennych w zestawie danych, które mogą korelować z widocznością.

  - **Interpolacja**: Jeśli dane są szeregami czasowymi, interpolacja może być dobrą metodą na oszacowanie brakujących wartości. Możesz użyć liniowej interpolacji, która wypełni brakujące wartości na podstawie wartości sąsiadujących w czasie, lub zastosować bardziej zaawansowane metody interpolacji, jak interpolacja splajnowa.

  - **Uzupełnienie na podstawie grupowania**: Jeśli brakujące wartości visibility mogą być różne dla różnych warunków pogodowych lub innych zmiennych, możesz uzupełnić brakujące dane średnią lub medianą dla odpowiednich grup. Na przykład, jeśli widoczność jest zwykle niższa podczas opadów deszczu, możesz obliczyć średnią widoczność dla dni deszczowych i dni bez deszczu oddzielnie i użyć tych wartości do wypełnienia braków.

  - **Zastosowanie wartości z najbliższego sąsiedztwa (k-nearest neighbors - KNN)**: Metoda ta polega na znajdowaniu 'k' najbliższych sąsiadów danej obserwacji z brakującymi danymi i wypełnieniu braku na podstawie wartości widoczności tych sąsiadów.

  *Wniosek:*
  
  *- Na potrzeby naszego projektu użyto metody imputacji średniej.*


# Obliczenie średniej dla kolumny 'visibility'
mean_visibility = df_weather['visibility'].mean()

# Uzupełnienie brakujących wartości w kolumnie 'visibility' średnią
df_weather['visibility'].fillna(mean_visibility, inplace=True)

# Usunięcie kolumn z pustymi danymi
columns_to_drop = ['sea_level', 'grnd_level', 'wind_gust', 'rain_1h', 'rain_3h', 'snow_1h', 'snow_3h']
df_weather.drop(columns=columns_to_drop, inplace=True)

#Ponowne sprawdzenie pustych wartosci
missing_data = df_weather.isnull().sum()
missing_data

# Wygenerowanie podsumowania statystycznego dla DataFrame df_weather
df_weather.describe()

# Usunięcie "+0000 UTC" z końca stringów w kolumnie 'dt_iso' i konwersja do formatu datetime
df_weather['dt_iso'] = pd.to_datetime(df_weather['dt_iso'].str.replace(r' \+\d{4} UTC', '', regex=True))

Wyrażenie regularne r' \+\d{4} UTC'
1.  : Znak spacji. Oznacza, że wyrażenie regularne zaczyna się od pojedynczej spacji.
2. \+: Znak plus (+). Ponieważ plus jest specjalnym znakiem w wyrażeniach regularnych, trzeba go poprzedzić backslashem (\), aby oznaczał dosłownie znak plus.
3. \d{4}: Oznacza cztery cyfry. \d jest skrótem dla dowolnej cyfry (0-9), a {4} oznacza, że muszą wystąpić dokładnie cztery cyfry.
4. UTC: Dosłowny ciąg znaków " UTC". Po czterech cyfrach musi wystąpić dokładnie ten ciąg znaków.

# Wyświetlenie pierwszych pięciu wierszy DataFrame df_weather
df_weather.head()

# Konwersja kolumny 'dt_iso' do formatu datetime bez informacji o strefie czasowej
df_weather['dt_iso'] = pd.to_datetime(df_weather['dt_iso'], format='%Y-%m-%d %H:%M:%S', errors='coerce')

# Wyświetlenie pierwszych pięciu wierszy DataFrame df_weather
df_weather.head()

# Przekształcenie kolumny 'dt_iso' na typ daty, zignorowanie czasu
df_weather['date'] = pd.to_datetime(df_weather['dt_iso']).dt.date

# Grupowanie danych po dacie i mieście
daily_weather = df_weather.groupby(['date', 'city_name']).agg({
    'lat': 'first',  # Pobranie pierwszej wartości szerokości geograficznej dla każdej grupy (zakładając, że się nie zmienia)
    'lon': 'first',  # Pobranie pierwszej wartości długości geograficznej dla każdej grupy (zakładając, że się nie zmienia)
    'temp': 'mean',  # Obliczenie średniej temperatury dla każdej grupy
    'feels_like': 'mean',  # Obliczenie średniej odczuwalnej temperatury dla każdej grupy
    'temp_min': 'min',  # Pobranie minimalnej temperatury dla każdej grupy
    'temp_max': 'max',  # Pobranie maksymalnej temperatury dla każdej grupy
    'pressure': 'mean',  # Obliczenie średniego ciśnienia dla każdej grupy
    'humidity': 'mean',  # Obliczenie średniej wilgotności dla każdej grupy
    'wind_speed': 'mean',  # Obliczenie średniej prędkości wiatru dla każdej grupy
    'clouds_all': 'mean',  # Obliczenie średniego zachmurzenia dla każdej grupy
    'weather_main': lambda x: x.mode()[0] if not x.mode().empty else 'Unknown'  # Wybranie najczęściej występującego opisu pogody, a jeśli brak trybu, zwrócenie 'Unknown'
}).reset_index()  # Zresetowanie indeksu, aby uzyskać płaski DataFrame


# Funkcja adjust_weather służy do dostosowania wartości w kolumnie 'weather_main' na podstawie wartości w kolumnie 'clouds_all'.
# Głównym celem tej funkcji jest przypisanie bardziej ogólnych kategorii pogodowych, takich jak 'Cloudy' lub 'Clear', w zależności od stopnia zachmurzenia, które reprezentowane jest przez wartość w kolumnie 'clouds_all'.
def adjust_weather(row):
    if row['clouds_all'] >= 80: # Jeśli wartość w kolumnie 'clouds_all' jest większa lub równa 80, uznajemy, że jest to zachmurzone niebo, więc przypisujemy 'Cloudy'
        return 'Cloudy'
    elif row['clouds_all'] <= 20: # Jeśli wartość w kolumnie 'clouds_all' jest mniejsza lub równa 20, uznajemy, że jest to bezchmurne niebo, więc przypisujemy 'Clear'
        return 'Clear'
    else:
        return row['weather_main'] # W przeciwnym razie, gdy wartość w kolumnie 'clouds_all' znajduje się między 20 a 80, zwracamy aktualną wartość z kolumny 'weather_main'
    
# Zastosowanie funkcji adjust_weather dla każdego wiersza w kolumnie 'weather_main'
daily_weather['weather_main'] = daily_weather.apply(adjust_weather, axis=1)

<u>*_Uwaga:_*</u>

Kod ten agreguje dane w kolumnie weather_main dla grupowanego zestawu danych i wybiera najczęściej występującą wartość (modę) dla tej kolumny. Funkcja mode() z pakietu pandas zwraca najczęściej występującą wartość w danym zbiorze danych. W kontekście tego kodu:

- x.mode()[0]: Zwraca pierwszą najczęściej występującą wartość z kolumny weather_main dla danego dnia i miasta. Ponieważ mode() może zwrócić kilka wartości (w przypadku, gdy więcej niż jeden wynik ma tę samą maksymalną liczbę wystąpień), indeks [0] gwarantuje, że zostanie wybrana tylko jedna z nich.

- if not x.mode().empty: Sprawdza, czy wynik funkcji mode() nie jest pusty. Funkcja mode() może zwrócić pusty wynik, jeśli stosowana jest na pustym zbiorze danych, co może się zdarzyć, gdy wszystkie dane dla danej grupy są NaN lub grupa jest pusta.

- else 'Unknown': Jeśli nie ma żadnej mody (czyli x.mode().empty jest prawdziwe), funkcja zwraca 'Unknown'. To zapewnia wartość domyślną, w przypadku gdy nie można określić typowej pogody dla danej grupy.

- Kod ten jest użyteczny w sytuacjach, gdy chcemy zredukować zmienność danych pogodowych do jednej reprezentatywnej wartości dla każdego dnia, co jest przydatne w analizach związanych z prognozowaniem energii czy badaniach nad wpływem warunków atmosferycznych na różne zjawiska.

# Wyświetlenie pierwszych pięciu wierszy DataFrame df_weather
daily_weather.head() 

# Wyświetlenie unikalnych wartości w kolumnie 'weather_main' DataFrame daily_weather
daily_weather["weather_main"].unique()

# Wyświetlenie pierwszych pięciu wierszy DataFrame daily_weather, gdzie wartość w kolumnie 'weather_main' wynosi 'Mist'
daily_weather[daily_weather["weather_main"]=="Mist"].head()

# Zapisanie DataFrame daily_weather do pliku CSV
daily_weather.to_csv('../2. Data/processed/combined_data_processed.csv')
